assignment2/requirements.txt
openai
chromadb
python-dotenv
colorama
langchain_text_splitters
wikipedia
tqdm
